Challenge 3
Added a very simple score below the word list. The score is computed directly from the used words array and is nothing more than a count of the words and the sum of their lengths. Also did a little bit of tidy up in the code by moving the view controls into computed properties.